💾Restructured the character methods in Core
