+++ Smash_Combos.Core/Cqrs/Characters/CharacterProfile.cs
@@ -1,0 +1,16 @@
+﻿using AutoMapper;
+using Smash_Combos.Domain.Models;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Smash_Combos.Core.Cqrs.Characters
+{
+    public class CharacterProfile : Profile
+    {
+        public CharacterProfile()
+        {
+            CreateMap<Character, CharacterResponse>();
+        }
+    }
+}
