+++ Smash_Combos.Core/Cqrs/Characters/GetCharacter/GetCharacterRequest.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace Smash_Combos.Core.Cqrs.Characters.GetCharacter
 {
-    public class GetCharacterRequest : IRequest<GetCharacterResponse>
+    public class GetCharacterRequest : IRequest<CharacterResponse>
     {
         public string VariableName { get; set; }
     }
