+++ Smash_Combos.Core/Cqrs/Characters/GetCharacters/GetCharactersRequest.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace Smash_Combos.Core.Cqrs.Characters.GetCharacters
 {
-    public class GetCharactersRequest : IRequest<IEnumerable<GetCharactersResponse>>
+    public class GetCharactersRequest : IRequest<IEnumerable<CharacterResponse>>
     {
         public string Filter { get; set; }
     }
